extract enterprise frontend to shared package for external invocation
This is a copy-and-paste of the cmd/enterprise/frontend enterprise service constructors to a shared package so that the enterprise frontend can be invoked from a different entrypoint. There are no logic changes.
